separating the comparison code
